+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TemporalFieldValue.java
@@ -29,7 +29,7 @@ final class TemporalFieldValue<T extends Temporal> extends AbstractComponentValu
 	TemporalFieldValue(TemporalField<T> component, UpdateOn updateOn) {
 		super(component);
 		if (updateOn == UpdateOn.VALUE_CHANGE) {
-			component.addListener(value -> notifyListeners());
+			component.valueObserver().addListener(this::notifyListeners);
 		}
 		else {
 			component.addFocusListener(new NotifyOnFocusLost());