+++ changelog.md
@@ -19,6 +19,8 @@ Codion Change Log
 ### is.codion.swing.common.model
 - FilteredTableSortModel, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredTableSearchModel, add/removeListener methods replaced with methods that return the underlying event observer.
+### is.codion.swing.common.ui
+- TemporalField, add/removeListener methods replaced with methods that return the underlying event observer.
 
 ## 0.17.34
 ### is.codion.common.core